[python] Handle returning values in direct calls and cudaq::run (#2881)
* * Added tests for arrays and tuples
* Added support for 'Ordered' schema
* WIP: Labeled schema


* * WIP: Support for labeld arrays. Clean-up required.


* * Added logic to parse labeled tuples
* Refactored the if-else logic for data type check into a look up map
* Free buffer in the tests


* * Pass strings by ref
* Remove redundant blank lines


* Update cudaq.run for python


* Cleanup


* Cleanup


* * Move header file inclusion to make gcc-12 compliant.


* Address CR comments


* Add more tests and fixes


* Add more tests and fixes


* Cleanup


* Handle returning data class objects


* Merge with features/run


* Comment out failing test


* Comment out failing test


* Temp


* Temp


* Support returning values via extra return argument


* Add tests and fix issues


* Fix failing tests


* Fixed aligment issues and added tests


* Cleanup


* Detect tuple element types on when passing arguments to direct calls


* Fix an issue when reading a struct element


* Added tests


* Uodate comments


* Update comments


* Update test TODO comments


---------

+++ python/utils/OpaqueArguments.h
@@ -233,7 +233,7 @@ inline void handleStructMemberVariable(void *data, std::size_t offset,
           appendValue(data, (bool)value.cast<py::bool_>(), offset);
           return;
         }
-        appendValue(data, (std::size_t)value.cast<py::int_>(), offset);
+        appendValue(data, (std::int64_t)value.cast<py::int_>(), offset);
       })
       .Case([&](mlir::Float64Type ty) {
         appendValue(data, (double)value.cast<py::float_>(), offset);
